+++ R/FilterVariance.R
@@ -1,16 +1,6 @@
 #' @title Variance Filter
 #'
-#' @usage NULL
 #' @name mlr_filters_variance
-#' @format [R6::R6Class] inheriting from [Filter].
-#' @include Filter.R
-#'
-#' @section Construction:
-#' ```
-#' FilterVariance$new()
-#' mlr_filters$get("variance")
-#' flt("variance")
-#' ```
 #'
 #' @description Variance filter calling [stats::var()].
 #'
@@ -26,21 +16,46 @@
 #' head(filter$scores, 3)
 #' as.data.table(filter)
 FilterVariance = R6Class("FilterVariance", inherit = Filter,
+
   public = list(
-    initialize = function() {
+
+    #' @description Create a FilterVariance object.
+    #' @param id (`character(1)`)\cr
+    #'   Identifier for the filter.
+    #' @param task_type (`character()`)\cr
+    #'   Types of the task the filter can operator on. E.g., `"classif"` or
+    #'   `"regr"`.
+    #' @param param_set ([paradox::ParamSet])\cr
+    #'   Set of hyperparameters.
+    #' @param feature_types (`character()`)\cr
+    #'   Feature types the filter operates on.
+    #'   Must be a subset of
+    #'   [`mlr_reflections$task_feature_types`][mlr3::mlr_reflections].
+    #' @param packages (`character()`)\cr
+    #'   Set of required packages.
+    #'   Note that these packages will be loaded via [requireNamespace()], and
+    #'   are not attached.
+    initialize = function(id = "variance",
+      task_type = c("classif", "regr"),
+      param_set = ParamSet$new(list(
+        ParamLgl$new("na.rm", default = TRUE)
+      )),
+      packages = "stats",
+      feature_types = c("integer", "numeric")) {
       super$initialize(
-        id = "variance",
-        packages = "stats",
-        feature_types = c("integer", "numeric"),
-        task_type = c("classif", "regr"),
-        param_set = ParamSet$new(list(
-          ParamLgl$new("na.rm", default = TRUE)
-        ))
+        id = id,
+        task_type = task_type,
+        param_set = param_set,
+        feature_types = feature_types,
+        packages = packages
       )
       self$param_set$values = list(na.rm = TRUE)
-    },
+    }
+  ),
 
-    calculate_internal = function(task, nfeat) {
+  private = list(
+
+    .calculate = function(task, nfeat) {
       na.rm = self$param_set$values$na.rm %??% TRUE
       map_dbl(task$data(cols = task$feature_names), var, na.rm = na.rm)
     }