+++ man-roxygen/field_feature_types.R
@@ -1,0 +1,2 @@
+#' @field feature_types (`character()`)\cr
+#'   Feature types of the filter.