+++ man-roxygen/field_packages.R
@@ -1,0 +1,2 @@
+#' @field packages ([character()])\cr
+#'   Packages which this filter is relying on.