+++ man-roxygen/field_task_properties.R
@@ -1,0 +1,2 @@
+#' @field task_properties (`character()`)\cr
+#'   [mlr3::Task]task properties.