+++ NEWS.md
@@ -1,3 +1,9 @@
+# mlr3filters 0.1.1.9002
+
+* Use `private$.calculate` instead of public calculate method for Filters 
+* rename Filter ID of `FilterFindCorrelation` from `findCorrelation` to `find_correlation`
+
+
 # mlr3filters 0.1.1.9001
 
 * new filter `FilterFindCorrelation` (#62, @mb706)